Cleaned up unused comments & unnecessary returns. Separated p2p & runtime modules into their own files under modules package.
